+++ src/views/schedule-details/__fixtures__/schedule-details-runs-chart-popover.ts
