+++ src/views/schedule-details/schedule-details-runs-chart-popover-trigger/schedule-details-runs-chart-popover-trigger.types.ts
@@ -1,4 +1,4 @@
-import { type ChartRunPopoverEntry } from '@/views/schedule-details/schedule-details-runs-chart-run-popover/schedule-details-runs-chart-run-popover.types';
+import { type ChartRunPopoverEntry } from '@/views/schedule-details/schedule-details-runs-chart-popover/schedule-details-runs-chart-popover.types';
 
 export type Props = {
   x: number;
